requirements.txt
addict==2.4.0
arrow==1.3.0
asttokens==3.0.0
attrs==25.3.0
backcall==0.2.0
blinker==1.8.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
comm==0.2.2
ConfigArgParse==1.7.1
contourpy==1.1.1
cycler==0.12.1
dash==3.1.1
decorator==5.2.1
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.13.1
Flask==3.0.3
fonttools==4.57.0
fsspec==2024.6.1
idna==3.10
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipython==8.12.3
ipywidgets==8.1.7
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
kiwisolver==1.4.7
lazy_loader==0.4
llvmlite==0.36.0
MarkupSafe==2.1.5
matplotlib==3.3.4
matplotlib-inline==0.1.7
mpmath==1.3.0
narwhals==1.42.1
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.1
numba==0.53.1
numpy==1.19.0
nvidia-cublas-cu12==12.4.2.65
nvidia-cuda-cupti-cu12==12.4.99
nvidia-cuda-nvrtc-cu12==12.4.99
nvidia-cuda-runtime-cu12==12.4.99
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.0.44
nvidia-curand-cu12==10.3.5.119
nvidia-cusolver-cu12==11.6.0.99
nvidia-cusparse-cu12==12.3.0.142
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.4.99
open3d==0.19.0
packaging==25.0
pandas==1.1.5
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
plotly==6.2.0
progressbar2==4.5.0
prompt_toolkit==3.0.51
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pykitti==0.3.1
pyparsing==3.1.4
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
python-utils==3.8.2
pytz==2025.2
PyWavelets==1.4.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.4
retrying==1.4.0
rpds-py==0.20.1
scikit-image==0.17.2
scikit-learn==1.3.2
scipy==1.5.4
six==1.17.0
stack-data==0.6.3
sympy==1.13.3
threadpoolctl==3.5.0
tifffile==2023.7.10
tqdm==4.67.1
traitlets==5.14.3
transforms3d==0.4.2
triton==3.0.0
types-python-dateutil==2.9.0.20241206
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.0.6
widgetsnbextension==4.0.14
zipp==3.20.2